Commit Refactor Double to BigDecimal

+++ src/main/java/ru/digitos/TransactionProcessor.java
@@ -1,9 +1,10 @@
 package ru.digitos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class TransactionProcessor {
-    void processTransaction(List<BankAccount> accounts, double amount) {
+    void processTransaction(List<BankAccount> accounts, BigDecimal amount){
         for (BankAccount account : accounts) {
             account.withdraw(amount);
         }
